Fix "Page required to force-reload when signing in" issue
