+++ src/Client/Pages/Authentication/Login.razor.cs
@@ -1,9 +1,9 @@
-﻿using BlazorHero.CleanArchitecture.Application.Requests.Identity;
+﻿using Blazored.FluentValidation;
+using BlazorHero.CleanArchitecture.Application.Requests.Identity;
 using Microsoft.AspNetCore.Components.Authorization;
 using MudBlazor;
 using System.Security.Claims;
 using System.Threading.Tasks;
-using Blazored.FluentValidation;
 
 namespace BlazorHero.CleanArchitecture.Client.Pages.Authentication
 {
@@ -25,12 +25,7 @@ namespace BlazorHero.CleanArchitecture.Client.Pages.Authentication
         private async Task SubmitAsync()
         {
             var result = await _authenticationManager.Login(_tokenModel);
-            if (result.Succeeded)
-            {
-                _snackBar.Add(string.Format(_localizer["Welcome {0}"], _tokenModel.Email), Severity.Success);
-                _navigationManager.NavigateTo("/", true);
-            }
-            else
+            if (!result.Succeeded)
             {
                 foreach (var message in result.Messages)
                 {
@@ -45,7 +40,7 @@ namespace BlazorHero.CleanArchitecture.Client.Pages.Authentication
 
         void TogglePasswordVisibility()
         {
-            if(_passwordVisibility)
+            if (_passwordVisibility)
             {
                 _passwordVisibility = false;
                 _passwordInputIcon = Icons.Material.Filled.VisibilityOff;
